+++ src/KeyUnit.cpp
@@ -273,7 +273,7 @@ void KeyUnit::removeKey( TKey * pT )
 }
 
 
-qint64 KeyUnit::getNewID()
+int KeyUnit::getNewID()
 {
     return ++mMaxID;
 }
