+++ src/TTimer.h
@@ -72,8 +72,8 @@ public:
     void             disableTimer();
     void             enableTimer( QString & );
     void             disableTimer( QString & );
-    void             enableTimer( qint64 );
-    void             disableTimer( qint64 );
+    void             enableTimer( int );
+    void             disableTimer( int );
     void             killTimer();
 
     bool             isOffsetTimer();
